[BUG] Character set and collate must occur in data type definition
Resolves issue where an SQL error occurs when defining a character set,
collation, and default value while adding a column to a table. As
character set and collation must occur in the data type definition,
prior to any other item in the column definition, such as default,
auto_increment or comment.

MySQL 5.0 documentation for the syntax of create table can be found at
http://dev.mysql.com/doc/refman/5.0/en/create-table.html