[FEATURE] Added configuration value to specify environment migration directory name
Added optional database environment configuration value to set the name
of the directory that contains migrations to be run on that
environment, where the name of the directory is not equal to the name
of the database.

Where the directory key is not set, database value will be used instead.

Eg. If development database migrations were stored in a directory
within migrations_dir called 'trunk', I would add the following key
value pair to the development array in config.

'directory' => 'trunk'