Rename db:schema to db:structure and replaced db:schema to also dump the migration versions
To be able to really use migrations in a way that makes sense, you need to know in what version the
database was when you made the schema dump so anyone can run a copy of the output and be up and
running. Otherwise, the database doesn't know in what version it is and will always try to run
all the migrations again which isn't working well.

For those that still only want the structure, there is now a db:structure command instead.